change loop counters and boundaries from size_t (64 bits) to uint32_t (32 bits)

+++ HLTrigger/Timer/plugins/FastTimerServiceClient.cc
@@ -89,8 +89,8 @@ FastTimerServiceClient::endRun(edm::Run const & run, edm::EventSetup const & set
   TH1F * path_total      = dqm->get(m_dqm_path + "/path_total_time")->getTH1F();
   TH1F * path_exclusive  = dqm->get(m_dqm_path + "/path_exclusive_time")->getTH1F();
 
-  size_t size = path_total->GetXaxis()->GetNbins();
-  for (size_t i = 0; i < size; ++i) {
+  uint32_t size = path_total->GetXaxis()->GetNbins();
+  for (uint32_t i = 0; i < size; ++i) {
     // extract the list of Paths and EndPaths from the bin labels of "path_total_time"
     std::string label = path_total->GetXaxis()->GetBinLabel(i+1);   // bin count from 1 (bin 0 is underflow)
     if (( me = dqm->get(m_dqm_path + "/Paths/" + label + "_total") ))
@@ -106,7 +106,7 @@ FastTimerServiceClient::endRun(edm::Run const & run, edm::EventSetup const & set
   //  - the running time spent in each module (total time spent in that module, averaged over the events where that module actually ran)
   //  - the "efficiency" of each module (number of time a module succeded divided by the number of times the has run)
   dqm->setCurrentFolder(m_dqm_path + "/Paths");
-  for (size_t p = 1; p <= size; ++p) {
+  for (uint32_t p = 1; p <= size; ++p) {
     // extract the list of Paths and EndPaths from the bin labels of "path_total_time"
     std::string label = path_total->GetXaxis()->GetBinLabel(p);
     MonitorElement * me_counter = dqm->get( m_dqm_path + "/Paths/" + label + "_module_counter" );
@@ -115,13 +115,13 @@ FastTimerServiceClient::endRun(edm::Run const & run, edm::EventSetup const & set
       continue;
     TH1F * counter = me_counter->getTH1F();
     TH1F * total   = me_total  ->getTH1F();
-    size_t bins = counter->GetXaxis()->GetNbins();
+    uint32_t bins = counter->GetXaxis()->GetNbins();
     double min  = counter->GetXaxis()->GetXmin();
     double max  = counter->GetXaxis()->GetXmax();
     TH1F * average    = dqm->book1D(label + "_module_average",    label + " module average",    bins, min, max)->getTH1F();
     TH1F * running    = dqm->book1D(label + "_module_running",    label + " module running",    bins, min, max)->getTH1F();
     TH1F * efficiency = dqm->book1D(label + "_module_efficiency", label + " module efficiency", bins, min, max)->getTH1F();
-    for (size_t i = 1; i <= bins; ++i) {
+    for (uint32_t i = 1; i <= bins; ++i) {
       const char * module = counter->GetXaxis()->GetBinLabel(i);
       average   ->GetXaxis()->SetBinLabel(i, module);
       running   ->GetXaxis()->SetBinLabel(i, module);
